functions/server/requirements.txt
asgiref==3.4.1
attrs==21.2.0
CacheControl==0.12.6
cachetools==4.2.2
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
fastapi==0.62.0
firebase-admin==4.5.3
google-api-core==1.31.2
google-api-python-client==2.20.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-cloud-appengine-logging==0.1.4
google-cloud-audit-log==0.1.1
google-cloud-core==1.6.0
google-cloud-firestore==2.3.1
google-cloud-logging==2.6.0
google-cloud-storage==1.37.0
google-crc32c==1.1.5
google-resumable-media==1.3.3
googleapis-common-protos==1.53.0
grpcio==1.40.0
h11==0.12.0
httplib2==0.19.1
httptools==0.2.0
idna==2.10
iniconfig==1.1.1
jmespath==0.10.0
msgpack==1.0.2
packaging==20.9
pluggy==0.13.1
proto-plus==1.19.0
protobuf==3.17.3
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.6.1
pyparsing==2.4.7
pytest==6.2.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-multipart==0.0.5
pytz==2021.1
PyYAML==5.3.1
requests==2.25.0
rsa==4.5
six==1.15.0
starlette==0.13.6
toml==0.10.2
uritemplate==3.0.1
urllib3==1.25.10
uvicorn==0.12.3
uvloop==0.16.0
watchgod==0.6
websocket-client==1.1.0
websockets==8.1